refactor: storeHero에서 mainInfo 분리
